feat: add category, question, exam and test command service impl

+++ src/main/java/pe/edu/upc/MLTDAH/tests/infrastructure/persistence/jpa/TestRepository.java
@@ -14,6 +14,6 @@ public interface TestRepository extends JpaRepository<Test, Long> {
     List<Test> findAll();
     List<Test> findAllByStudentId(Long studentId);
     @Query("SELECT t FROM Test t WHERE t.studentId = :studentId AND FUNCTION('YEAR', t.createdAt) = :year")
-    List<Test> findByStudentIdAndYear(@Param("studentId") Long studentId, @Param("year") int year);
+    List<Test> findByStudentIdAndYear(@Param("studentId") Long studentId, @Param("year") Long year);
     Optional<Test> findById(Long id);
 }
